requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
waitress>=3.0.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
edge-tts>=6.1.0
Pillow>=10.0.0
langdetect==1.0.9
